Adjusted javascript to handle errors for favorite list request more 'robust'
At least now it shows the error to you gracefully, though why the node.js server has this problem was unclear (occasionally it would not handle the request, and throw an error). After a restart of the node.js server it seemed to stop doing this, but I implemented this error handling in case it happens again.